requirements.txt
# Requirements
numpy
torch>=1.8
gpytorch
faiss-cpu
# setuptools_scm[toml]
# setuptools_scm_git_archive
